requirements.txt
beautifulsoup4==4.12.3
openai==1.56.2
python-dotenv==1.0.1
pytube==15.0.0
requests==2.32.3
youtube_transcript_api==0.6.2